Create Toffee_problem.cpp
first we create an vector array and push those price of toffee which is less than or equal to the total price(m) then after that sort the array. now sum the element of array from first index and check the sum is greator than or equal to m and increament a count .if sum is equal or greator then break the while loop. and print count as final output.